Created history callbacks and refactor similar callback for homepage and history
