+++ app/callbacks/data_callbacks.py
@@ -3,13 +3,12 @@
 # This program is licensed under the Apache License 2.0.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
 
-import json
 from io import StringIO
 
 import dash
 import logging_config
 import pandas as pd
-from dash import callback_context, dcc, html
+from dash import dcc, html
 from dash.dependencies import Input, Output, State
 from dash.exceptions import PreventUpdate
 from main import app
@@ -17,7 +16,7 @@ from main import app
 import config as cfg
 from pages.cameras_status import display_cam_cards
 from services import api_client, get_token
-from utils.data import process_bbox
+from utils.data import load_detections
 
 logger = logging_config.configure_logging(cfg.DEBUG, cfg.SENTRY_DSN)
 
@@ -225,63 +224,25 @@ def api_watcher(n_intervals, api_cameras, local_sequences, user_token):
     State("are_detections_loaded", "data"),
     prevent_initial_call=True,
 )
-def load_detections(api_sequences, sequence_id_on_display, api_detections, are_detections_loaded):
-    # Deserialize data
-    api_sequences = pd.read_json(StringIO(api_sequences), orient="split")
-    if api_sequences.empty:
-        return dash.no_update, dash.no_update, dash.no_update
+def load_detections_homepage(api_sequences, sequence_id_on_display, api_detections, are_detections_loaded):
+    return load_detections(api_sequences, sequence_id_on_display, api_detections, are_detections_loaded)
 
-    sequence_id_on_display = str(sequence_id_on_display)
-    are_detections_loaded = json.loads(are_detections_loaded)
-    api_detections = json.loads(api_detections)
 
-    # Initialize sequence_on_display
-    sequence_on_display = pd.DataFrame().to_json(orient="split")
-
-    # Identify which input triggered the callback
-    ctx = callback_context
-    if not ctx.triggered:
-        raise PreventUpdate
-
-    triggered_input = ctx.triggered[0]["prop_id"].split(".")[0]
-
-    if triggered_input == "sequence_id_on_display":
-        # If the displayed sequence changes, load its detections if not already loaded
-        if sequence_id_on_display not in api_detections:
-            response = api_client.fetch_sequences_detections(sequence_id_on_display)
-            detections = pd.DataFrame(response.json())
-            if not detections.empty and "bboxes" in detections.columns:
-                detections["processed_bboxes"] = detections["bboxes"].apply(process_bbox)
-            api_detections[sequence_id_on_display] = detections.to_json(orient="split")
-
-        sequence_on_display = api_detections[sequence_id_on_display]
-        last_seen_at = api_sequences.loc[
-            api_sequences["id"].astype("str") == sequence_id_on_display, "last_seen_at"
-        ].iloc[0]
-
-        # Ensure last_seen_at is stored as a string
-        are_detections_loaded[sequence_id_on_display] = str(last_seen_at)
-
-    else:
-        # If no specific sequence is triggered, load detections for the first missing sequence
-        for _, row in api_sequences.iterrows():
-            sequence_id = str(row["id"])
-            last_seen_at = row["last_seen_at"]
-
-            if sequence_id not in are_detections_loaded or are_detections_loaded[sequence_id] != str(last_seen_at):
-                response = api_client.fetch_sequences_detections(sequence_id)
-                detections = pd.DataFrame(response.json())
-                if not detections.empty and "bboxes" in detections.columns:
-                    detections["processed_bboxes"] = detections["bboxes"].apply(process_bbox)
-                api_detections[sequence_id] = detections.to_json(orient="split")
-                are_detections_loaded[sequence_id] = str(last_seen_at)
-                break
-
-        # Clean up old sequences that are no longer in api_sequences
-        sequences_in_api = api_sequences["id"].astype("str").values
-        to_drop = [key for key in are_detections_loaded if key not in sequences_in_api]
-        for key in to_drop:
-            are_detections_loaded.pop(key, None)
-
-    # Serialize and return data
-    return json.dumps(are_detections_loaded), sequence_on_display, json.dumps(api_detections)
+@app.callback(
+    [
+        Output("are_detections_loaded_history", "data"),
+        Output("sequence_on_display_history", "data"),
+        Output("api_detections_history", "data"),
+    ],
+    [
+        Input("api_sequences_history", "data"),
+        Input("sequence_id_on_display_history", "data"),
+        Input("api_detections_history", "data"),
+    ],
+    State("are_detections_loaded_history", "data"),
+    prevent_initial_call=True,
+)
+def load_detections_history(api_sequences, sequence_id_on_display, api_detections, are_detections_loaded):
+    return load_detections(
+        api_sequences, sequence_id_on_display, api_detections, are_detections_loaded, id_suffix="_history"
+    )
